y1oing-music_Bot/requirements.txt
# Discord Bot Framework
discord.py

# For loading environment variables from .env file
python-dotenv

# For fetching audio stream info from YouTube and other sites
yt-dlp

# Required by discord.py for voice functionality (encryption)
PyNaCl

# Manually specify a known stable version of aiohttp
aiohttp==3.9.1